requirements.txt
fastapi==0.110.2
pydantic==2.7.0
boto3==1.34.147
aiomysql==0.2.0
uvicorn==0.29.0
python-dotenv==1.0.1
passlib==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9